change "data gathering" toggle label on admin-overview.html
The term "data gathering" suggests telemetry or some kind of surveillance of users.

It wasn't until I checked the discussion section on Github that I understood this actually meant fetching data from 3rd party sources. This change is to change the English label for this toggle to avoid misunderstandings for users.